+++ src/bootstrap.php
@@ -47,3 +47,49 @@ if (!function_exists(__NAMESPACE__ . '\inContainer')) {
         return System::inContainer();
     }
 }
+
+if (!function_exists(__NAMESPACE__ . '\swap')) {
+    /**
+     * Swap variables values by references
+     *
+     * @param mixed $var1
+     * @param mixed $var2
+     * @return void
+     */
+    function swap(&$var1, &$var2): void
+    {
+        [$var1, $var2] = [$var2, $var1];
+    }
+}
+
+if (!function_exists(__NAMESPACE__ . '\swapTo')) {
+    /**
+     * Swap variables values to array
+     *
+     * @param mixed $var1
+     * @param mixed $var2
+     * @return array
+     */
+    function swapTo($var1, $var2): array
+    {
+        return [$var2, $var1];
+    }
+}
+
+if (!function_exists(__NAMESPACE__ . '\inRange')) {
+    /**
+     * Checks if int is in range
+     *
+     * @param int $value
+     * @param int $min
+     * @param int $max
+     * @return bool
+     */
+    function inRange(int $value, int $min, int $max): bool
+    {
+        if ($min > $max) {
+            swap($min, $max);
+        }
+        return ($min <= $value && $value <= $max);
+    }
+}