+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.4] - 2019-04-17
+### Added
+- function `swap()`
+- function `swapTo()`
+- function `inContainer()`
+- function `inRange()`
+
 ## [0.5.1] - 2019-03-15
 ### Added
 - class `Picklock::class` 
@@ -65,7 +72,8 @@ define('AlecRabbit\Helpers\Constants\I_01DAY', 86400);
 ## 0.0.16 - 2018-11-29
 
 
-[Unreleased]: https://github.com/alecrabbit/php-helpers/compare/0.5.1...HEAD
+[Unreleased]: https://github.com/alecrabbit/php-helpers/compare/0.5.4...HEAD
+[0.5.4]: https://github.com/alecrabbit/php-helpers/compare/0.5.1...0.5.4
 [0.5.1]: https://github.com/alecrabbit/php-helpers/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/alecrabbit/php-helpers/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/alecrabbit/php-helpers/compare/0.3.1...0.4.2